SS_prototype/requirements.txt
Flask
python-dotenv
google-generativeai
langchain-google-genai
langchain
chromadb
pypdf
langchain_community